requirements.txt
streamlit
tensorflow==2.21.0
numpy==2.2.6
pillow
ultralytics
opencv-python
